+++ .gitignore
@@ -8,3 +8,4 @@ bin/
 build/
 **swp
 **swo
+.idea
