--- src/main/java/com/jos/dem/solid/dip/PersonRepository.java
@@ -1,7 +1,0 @@
-package com.jos.dem.solid.dip;
-
-import java.util.List;
-
-public interface PersonRepository{
-  List<Person> findAll();
-}
