--- src/main/java/com/jos/dem/solid/dip/PersonService.java
@@ -1,17 +1,0 @@
-package com.jos.dem.solid.dip;
-
-import java.util.List;
-
-public class PersonService {
-
-  private PersonRepository personRepository;
-
-  public void setPersonRepository(PersonRepository personRepository){
-    this.personRepository = personRepository;
-  }
-
-  public List<Person> getAll(){
-    return personRepository.findAll();
-  }
-
-}
