--- src/main/java/com/jos/dem/solid/isp/Contractor.java
@@ -1,7 +1,0 @@
-package com.jos.dem.solid.isp;
-
-import java.math.BigDecimal;
-
-public interface Contractor extends Employee {
-  BigDecimal getBonus();
-}
