--- src/main/java/com/jos/dem/solid/isp/Employee.java
@@ -1,8 +1,0 @@
-package com.jos.dem.solid.isp;
-
-import java.math.BigDecimal;
-
-public interface Employee {
-  BigDecimal getBaseAmount();
-}
-
