--- src/main/java/com/jos/dem/solid/isp/Partner.java
@@ -1,7 +1,0 @@
-package com.jos.dem.solid.isp;
-
-import java.math.BigDecimal;
-
-public interface Partner extends Employee {
-  BigDecimal getProfits();
-}
