--- src/main/java/com/jos/dem/solid/lsp/Salesman.java
@@ -1,13 +1,0 @@
-package com.jos.dem.solid.lsp;
-
-import java.math.BigDecimal;
-
-public class Salesman implements Employee {
-	
-	private BigDecimal commission = new BigDecimal(10);
-
-	public BigDecimal getPaymentAmount() {
-		return monthlySalary.add(commission);
-  }
-
-}
