--- src/main/java/com/jos/dem/solid/ocp/Engineer.java
@@ -1,11 +1,0 @@
-package com.jos.dem.solid.ocp;
-
-import java.math.BigDecimal;
-
-public class Engineer implements Employee {
-	
-	public BigDecimal getPaymentAmount() {
-		return monthlySalary;
-	}
-
-}
