--- src/main/java/com/jos/dem/solid/ocp/Manager.java
@@ -1,13 +1,0 @@
-package com.jos.dem.solid.ocp;
-
-import java.math.BigDecimal;
-
-public class Manager implements Employee {
-	
-	private BigDecimal bonus = new BigDecimal(20);
-
-	public BigDecimal getPaymentAmount() {
-		return monthlySalary.add(bonus);
-	}
-
-}
