--- src/main/java/com/jos/dem/solid/srp/FuelPump.java
@@ -1,11 +1,0 @@
-package com.jos.dem.solid.srp;
-
-public class FuelPump {
-
-  public void reFuel(Car car){
-    while(!car.isFull()){
-      car.increment();
-    }
-  }
-
-}
