+++ src/main/java/com/josdem/solid/dip/Person.java
@@ -1,0 +1,15 @@
+package com.josdem.solid.dip;
+
+public class Person {
+
+  private String nickname;
+
+  public Person(String nickname){
+    this.nickname = nickname;
+  }
+
+  public String getNickname(){
+    return this.nickname;
+  }
+
+}
