+++ src/main/java/com/josdem/solid/dip/PersonRepository.java
@@ -1,0 +1,7 @@
+package com.josdem.solid.dip;
+
+import java.util.List;
+
+public interface PersonRepository{
+  List<Person> findAll();
+}
