+++ src/main/java/com/josdem/solid/isp/Contractor.java
@@ -1,0 +1,7 @@
+package com.josdem.solid.isp;
+
+import java.math.BigDecimal;
+
+public interface Contractor extends Employee {
+  BigDecimal getBonus();
+}
