+++ src/main/java/com/josdem/solid/isp/Employee.java
@@ -1,0 +1,8 @@
+package com.josdem.solid.isp;
+
+import java.math.BigDecimal;
+
+public interface Employee {
+  BigDecimal getBaseAmount();
+}
+
