+++ src/main/java/com/josdem/solid/isp/Partner.java
@@ -1,0 +1,7 @@
+package com.josdem.solid.isp;
+
+import java.math.BigDecimal;
+
+public interface Partner extends Employee {
+  BigDecimal getProfits();
+}
