+++ src/main/java/com/josdem/solid/lsp/Employee.java
@@ -1,0 +1,10 @@
+package com.josdem.solid.lsp;
+
+import java.math.BigDecimal;
+
+public interface Employee {
+	
+	final BigDecimal monthlySalary = new BigDecimal(100);
+	BigDecimal getPaymentAmount();
+
+}
