+++ src/main/java/com/josdem/solid/lsp/Engineer.java
@@ -1,0 +1,11 @@
+package com.josdem.solid.lsp;
+
+import java.math.BigDecimal;
+
+public class Engineer implements Employee {
+	
+	public BigDecimal getPaymentAmount() {
+		return monthlySalary;
+  }
+  
+}
