+++ src/main/java/com/josdem/solid/ocp/Employee.java
@@ -1,0 +1,10 @@
+package com.josdem.solid.ocp;
+
+import java.math.BigDecimal;
+
+public interface Employee {
+	
+	final BigDecimal monthlySalary = new BigDecimal(100);
+	BigDecimal getPaymentAmount();
+
+}
