+++ src/main/java/com/josdem/solid/ocp/Engineer.java
@@ -1,0 +1,11 @@
+package com.josdem.solid.ocp;
+
+import java.math.BigDecimal;
+
+public class Engineer implements Employee {
+	
+	public BigDecimal getPaymentAmount() {
+		return monthlySalary;
+	}
+
+}
