+++ src/main/java/com/josdem/solid/ocp/Manager.java
@@ -1,0 +1,13 @@
+package com.josdem.solid.ocp;
+
+import java.math.BigDecimal;
+
+public class Manager implements Employee {
+	
+	private BigDecimal bonus = new BigDecimal(20);
+
+	public BigDecimal getPaymentAmount() {
+		return monthlySalary.add(bonus);
+	}
+
+}
