+++ src/main/java/com/josdem/solid/srp/FuelPump.java
@@ -1,0 +1,11 @@
+package com.josdem.solid.srp;
+
+public class FuelPump {
+
+  public void reFuel(Car car){
+    while(!car.isFull()){
+      car.increment();
+    }
+  }
+
+}
