+++ src/test/groovy/com/jos/dem/solid/srp/CarSpec.groovy
@@ -1,12 +1,13 @@
 package com.jos.dem.solid.srp
 
+import com.josdem.solid.srp.Car
 import spock.lang.Specification
 
 class CarSpec extends Specification {
 
   void "should start with empty tank"(){
     when:'A car'
-      Car car = new Car()
+    Car car = new Car()
     then:'We expect is empty gas'  
       car.isEmpty() == true
   }
